requirements.txt
sentence-transformers
langchain-community
faiss-cpu
tqdm
google-genai
streamlit
pandas
numpy
